server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi[standard]>=0.118.0",
    "google-adk>=1.15.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
